(#7359) Fix bullet icons under relative_url_root
The relative path from the stylesheets directory to the images directory is
incorrect in tables.css. This problem has gone unnoticed because, when the app
is deployed at a root URL of /, both stylesheets/.. and stylesheets/../..
refer to the URL root. Fix the icons for relative_url_root deployments by
correcting the path.